pyproject.toml
[project]
name = "ethereum-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "duckdb-engine>=0.17.0",
    "eth-abi>=5.2.0",
    "eth-hash[pycryptodome]>=0.7.1",
    "eth-utils>=5.3.1",
    "fake-useragent>=2.2.0",
    "httpx-ws>=0.7.2",
    "httpx[http2]>=0.28.1",
    "kafka-python>=2.2.15",
    "loguru>=0.7.3",
    "orjson>=3.11.2",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    "uvloop>=0.21.0",
]
